+++ lib/ex_unit/lib/ex_unit/cli_formatter.ex
@@ -6,7 +6,8 @@ defmodule ExUnit.CLIFormatter do
   import ExUnit.Formatter, only: [format_time: 2, format_filters: 2, format_test_failure: 5, format_test_case_failure: 4]
 
   defrecord Config, tests_counter: 0, invalids_counter: 0, failures_counter: 0,
-                    skips_counter: 0, trace: false, color: true, previous: nil
+                    skips_counter: 0, trace: false, color: true, previous: nil,
+                    seed: nil
 
   ## Callbacks
 
@@ -120,6 +121,8 @@ defmodule ExUnit.CLIFormatter do
       config.invalids_counter > 0 -> IO.puts invalid(message, config)
       true                        -> IO.puts success(message, config)
     end
+
+    IO.puts "\nRandomized with seed #{config.seed}"
   end
 
   defp print_filters([include: include, exclude: exclude]) do